# Major speed-up of RGD scoring