This PR is about improving the R-group decomposition scoring speed by caching already settled matches rather than recomputing scores for all matches every time `process()` is called as currently happens.

The motivation for this PR is that I ran into an internal dataset of 567 molecules that was taking >10 minutes to process.
I analyzed our current RGD permutation scoring code, and I realized that since both the `Greedy` and `GreedyChunks` methods prune the "settled" match permutations to keep only the single best scoring permutation, there is no need to score all matches again and again every time `process()` is called, as their score will not vary anymore even though a new batch of molecules is added.
Therefore, I put some of the existing scoring logic into a new `RGroupScorer` class and implemented some caching logic in this class.
Then I made sure that, with the caching in place, I was still getting exactly the same results as I was getting from the current `master` branch.

The speed improvements are quite impressive, the more so the higher the degree of symmetry in the core(s), as symmetry makes the scoring step more time-consuming. Here follow some timings:
* the previously mentioned internal dataset (567 molecules) that used to take 11m7.673s now is processed in 2m11.138s (>5-fold speed-up)
* a public 1,4-DHP ChEMBL dataset that I assembled for benchmarking this PR (1069 1,4-dihydropyridines) that used to take 0m12.853s now is processed in 0m6.479s (~2-fold speed-up)
* a public bicyclo[2.2.2]octane ChEMBL dataset that I assembled for benchmarking this PR (1388 molecules) that used to take 233m51.679s now is processed in 40m58.195s (>5-fold speed-up)

See below some detailed statistics for the above three test cases, including the check of equality of PR results with the `master` branch results.

The Python benchmarking script:
[benchmark_rgd_scoring.py.gz](https://github.com/rdkit/rdkit/files/7206329/benchmark_rgd_scoring.py.gz)

The two public ChEMBL benchmark datasets:
[ChEMBL_bridged.smi.gz](https://github.com/rdkit/rdkit/files/7206333/ChEMBL_bridged.smi.gz)
[ChEMBL_1_4_DHP.smi.gz](https://github.com/rdkit/rdkit/files/7206331/ChEMBL_1_4_DHP.smi.gz)

# Internal 567-molecule dataset

### `master` branch

```
$ module purge; module load rdkit_master
$ time ( python rgd_perf.py )

real    11m14.800s
user    11m7.673s
sys     0m3.114s

base = "master"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)

fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134205803-0453e850-d22c-471d-b11e-4f80fe7939b2.png)
```
collected_results["master"] = (results, rejected)
```

### PR

```
$ module purge; module load rdkit_improve_rgd_scoring_speed
$ time ( python rgd_perf.py )

real    2m16.798s
user    2m11.138s
sys     0m3.024s


base = "improve_rgd_scoring_speed"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)
fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134206101-c2c767bc-8b56-4768-9c63-c338d31dc56c.png)
```
collected_results["PR"] = (results, rejected)
collected_results["master"] == collected_results["PR"]
True
len(collected_results["PR"][0]["Core"])
567
```

# ChEMBL 1,4-DHP 1069-molecule dataset

### `master` branch

```
$ module purge; module load rdkit_master
$ time ( dataset_scaffold='N1C=CCC=C1' dataset_base=ChEMBL_1_4_DHP python benchmark_rgd_scoring.py )

real    0m23.998s
user    0m12.853s
sys     0m1.398s

base = "ChEMBL_1_4_DHP_master/ChEMBL_1_4_DHP"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)

fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134245560-b607e131-90ad-47fc-828d-665b343af36c.png)
```
collected_results["master"] = (results, rejected)
```

### PR

```
$ module purge; module load rdkit_improve_rgd_scoring_speed
$ time ( dataset_scaffold='N1C=CCC=C1' dataset_base=ChEMBL_1_4_DHP python benchmark_rgd_scoring.py )

real    0m13.632s
user    0m6.479s
sys     0m1.298s

base = "ChEMBL_1_4_DHP_PR/ChEMBL_1_4_DHP"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)

fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134245791-892d3142-9e08-4e9d-b465-203a12342743.png)
```
collected_results["PR"] = (results, rejected)
collected_results["master"] == collected_results["PR"]
len(collected_results["master"][0]["Core"])
1069
```

# ChEMBL bicyclo[2.2.2]octane 1388-molecule dataset

### `master` branch

```
$ module purge; module load rdkit_master
$ time ( dataset_scaffold='C1CC2CCC1CC2' dataset_base=ChEMBL_bridged python benchmark_rgd_scoring.py )

real    234m43.518s
user    233m51.679s
sys     0m17.037s

base = "master/ChEMBL_bridged"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)

fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134246140-ec789979-fe9a-4132-a3c5-fb5700984f3e.png)
```
collected_results["master"] = (results, rejected)
```

### PR

```
$ module purge; module load rdkit_improve_rgd_scoring_speed
$ time ( dataset_scaffold='C1CC2CCC1CC2' dataset_base=ChEMBL_bridged python benchmark_rgd_scoring.py )

real    42m15.375s
user    40m58.195s
sys     1m6.836s

base = "ChEMBL_bridged_PR/ChEMBL_bridged"
with open(f"{base}_time_per_mol.pkl", "rb") as hnd:
    time_per_mol = pickle.load(hnd)
with open(f"{base}_results.pkl", "rb") as hnd:
    results = pickle.load(hnd)
with open(f"{base}_rejected.pkl", "rb") as hnd:
    rejected = pickle.load(hnd)

fig, ax = plt.subplots()
fig.set_size_inches(20, 10)
ax.plot(*zip(*time_per_mol))
plt.show()
```
![image](https://user-images.githubusercontent.com/5244385/134246316-9107f222-468e-4e49-9976-1ef5fb297399.png)
```
collected_results["PR"] = (results, rejected)
collected_results["master"] == collected_results["PR"]
True
len(collected_results["`master` branch"][0]["Core"])
1388
```