@ptosco 2021-09-22
The failure in the `Ubuntu_18_04_x64` build is unrelated to this PR (I think I have occasionally seen this before):
```
194/206 Test #194: pySimDivPickers ........................***Failed    0.40 sec
[22:11:49] the useCache argument is deprecated and ignored
...[22:11:49] the useCache argument is deprecated and ignored
[22:11:49] the useCache argument is deprecated and ignored
.[22:11:49] the useCache argument is deprecated and ignored
......F
======================================================================
FAIL: testNonUniqueCrash (__main__.TestCase)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/vsts/work/1/s/Code/SimDivPickers/Wrap/testPickers.py", line 139, in testNonUniqueCrash
    self.assertTrue(tuple(mm2) != tuple(mm1)) or (tuple(mm3) != tuple(mm1))
AssertionError: False is not true

----------------------------------------------------------------------
Ran 11 tests in 0.181s

FAILED (failures=1)
```