@greglandrum 2021-09-26
I forgot to comment on this in the original review: this kind of speedup is awesome to get and the way you're doing it seems really sensible to me.