@bp-kelley 2021-09-26
@ptosco @greglandrum for some reason, I'm not seeing the discussion/resolution on your comments above so the logic train is a bit confusing.