@ptosco 2021-09-27
@bp-kelley Hi Brian, I think I have addressed all your comments except unit tests for `RGroupScorer`.
I have duly checked that results haven’t changed compared to `master` (see above), and that should be sufficient.
Regarding testing the case where permutations exceed 1e6, yes, I have tested that case as part of the tests described above in the PR. Some of those datasets actually do incur into that situation and are therefore processed before `chunkSize` is reached and, yes, I have verified that the caching mechanism indeed works also in that case.