@bp-kelley 2021-09-30
@ptosco Nice, I have one more trick up my sleeve for a speed improvement and this foundation will help a ton I think.